requirements.txt
asgiref==3.2.3
bson==0.5.8
Cython==0.29.15
dataclasses==0.6
dj-database-url==0.5.0
Django==2.2.10
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
postgis==1.0.4
ppygis3==0.4
psycopg-postgis==0.1.3
psycopg2==2.8.4
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
whitenoise==5.0.1
